IOS-2588 Moved the code responsible for selecting a new model to MainCoordinator
