+++ Tangem/Modules/Auth/AuthRoutable.swift
@@ -11,7 +11,6 @@ import Foundation
 protocol AuthRoutable: AnyObject {
     func openOnboarding(with input: OnboardingInput)
     func openMain(with cardModel: CardViewModel)
-    func updateMain(with cardModel: CardViewModel)
     func openMail(with dataCollector: EmailDataCollector, recipient: String)
     func openDisclaimer(at url: URL, _ handler: @escaping (Bool) -> Void)
 }
